Remove `:format` default for `Gregorian#formats`
`:format` is not a valid `type` anyway, so it would never match anyway.

In the `ldml.dtd`:

* For `dateFormatLength`, `timeFormatLength` the `type` attribute is `#REQUIRED`.
* For `dateTimeFormatLength`, `type` attribute is `#IMPLIED`

There are no instances in the code where any of these are missing the `type` attribute.
So while in the future it could technically happen for `dateTimeFormatLength`,
I'd rather the code fail in that case, then silently pass with a nonsense value.